feature/network_selection (#35)
* allow devs to select network on building or deploying

* docker compose to start local node

* rename prepare to postinstall

* revert rename

* start/stop local development script

* fix prettier

* fix ganache-cli docker

* update docker-compose

* update

* build before test

* development artifacts to pass testing

* remove prepare

* ignore development artifacts

* update codegen

* update paths

* fix

* update

* remove prepare. needs a different approach

* revert

+++ contracts/interfaces/LiquidityPoolInterface.sol
@@ -2,14 +2,23 @@ pragma solidity ^0.5.8;
 
 interface LiquidityPoolInterface {
     // return 0 means not available for this trade
-    function getBidSpread(address fToken) external view returns (uint);
+    function getBidSpread(address fToken) external view returns (uint256);
     // return 0 means not available for this trade
-    function getAskSpread(address fToken) external view returns (uint);
+    function getAskSpread(address fToken) external view returns (uint256);
     // additional collaterla ratio this pool want to maintain. e.g. 0.5 ether means 150% of the issuing assets value
-    function getAdditionalCollateralRatio(address fToken) external view returns (uint);
+    function getAdditionalCollateralRatio(address fToken)
+        external
+        view
+        returns (uint256);
 
     // positive leverage means long, negative means short
-    function openPosition(address tradingPair, uint positionId, address quoteToken, int leverage, uint baseTokenAmount) external returns (bool);
+    function openPosition(
+        address tradingPair,
+        uint256 positionId,
+        address quoteToken,
+        int256 leverage,
+        uint256 baseTokenAmount
+    ) external returns (bool);
 
     event SpreadUpdated();
     event AdditoinalCollateralRatioUpdated();
